Deprecate ServerDetails class
We have recently introduced possibility to alter base_path for 
operations on the resource and shifted some of the services to use it. 
For compute.servers(details) we still use an inherited class with 
modified base_path. Remove this class and instead alter the base_path 
for the list servers operation. This also has an advantage, that we get 
a list of servers with details and are immediately able to execute 
operations on the instance without any conversion or re-fetch by id.

Change-Id: I4c4649ee1390676f3ccc6bf19e4b11b1964b5aaa