Replace openstack.org git:// URLs with https://
This is a mechanically generated change to replace openstack.org
git:// URLs with https:// equivalents.

This is in aid of a planned future move of the git hosting
infrastructure to a self-hosted instance of gitea (https://gitea.io),
which does not support the git wire protocol at this stage.

This review should result in no functional change.

Change-Id: If95a90e4110a5330535fc5b0357f181f27f3f918
Story: #2004627
Task: #29701