Do not disregard tags when updating stacks
At present, if tags are supplied to the update_stack function, it is
completely disregarded while create_stack in fact allows it. This leads
the Ansible module os_stack to also not be able to propagate tags
properly. This is suboptimal therefore this patch aims to sort the issue
out.

Resolves: https://github.com/ansible/ansible/pull/53757

Change-Id: I59e7c47b8c01f96ec561b8bd153804bac5f21800
Story: 2005226
Task: 30008