Split OpenStackCloud into reasonable pieces
With over 12k lines of code in openstackcloud class it is hardly 
manageable. Split methods on a service base into corresponding Mixings.

It still make sense to think how individual mixins are structured (it is 
not very nice, when function in an object use property of self, which is 
not defined in the class, but exists only in a resulting mixed object).
Initialization of individual Mixins in the connection should be also 
reviewed

Change-Id: I81050d66e3a2dc0997e3e7620ff2c4e1891bfeb7