Move set_temp_url_key logic into resource objects
These are resource-specific methods, so they should exist on the
Resource objects. Move the methods there and make the proxy methods
use them. Also, shift tests for them into the object_store tests.
This puts them more where they should be - and puts requests_mock
tests into those files, which is a good example for people.

While we're in there, stop creating a Proxy with a Session and use
the object_store Proxy that exists.

Change-Id: I4ba4456ff044c6c2b0c4ab5faeaf3e3cfd77194f