Remove and rearrange object bonged proxy tests
There are two test classes witha  setUp method and all of their
tests commented out. The commented tests pre-date the merge and
are implemented using httpretty, which is not what we use around
here. The fun part is though, since they subclass the proxy tests,
they are not adding any value - BUT cause the proxy tests to run
multiple times. Just remove them, since they're wasting energy.

Also, there are a couple of tests in their own classes, which is
similarly not the right design, as it causes all of the tests in
the parent class to be run again. Move them into the main class.

Change-Id: Ie328a3cd6e59cb9f40a57cd2c79bbd5890e1f2fe