Add support for vendor hooks
Add possibility to pass a hook in the vendor config, clouds-public.* or 
upon building a connection. This should be a string parameter - function
name to be executed. This gives possibility to register new services of 
the cloud automatically or alter behavior of the present services.
It would have not been necessary, if public clouds followed 
upstream-first aproach.

While we are here fix warnings on not closed files in the test_json

Change-Id: Ifd6c0847102af4f46e361dcb1a665829c77553b9