Support skipping unknown QP
In the cloud layer we receive multiple filters. We would like to pass as 
much filters as possible to the server to reduce response size. For this 
we need to know which QPs are supported by the given resource. Add 
support for passing everything given to the Resource.list function and 
silently skipping unknown parameters. This is not going to be a default 
behavior, but primarily used from cloud layer.

Change-Id: I53854ddcb86b92acdb64fabeb1de4ea4907c4a21