Support deleting all routes in update_router
There was previously no way to remove all routes from a router with the
update_router function because it only processed the routes if the
statement 'if routes' evaluated to True, which an empty list did not.

Modified the unit test for cloud/test_router so that it initialises the
test router with routes, and confirms that they are removed.

Change-Id: I9f31ad33eb500f59eb7634ea9e7e6502362a375b
Story: 2005089
Task: 29685