Add access alias (aka) for the resource attributes
In the compute.server we are currently exposing names, which are only 
"aliases" for known attributes (i.e. accessIPv4 vs access_ipv4). Since 
we want to drop the _normalization completely it is necessary to expose 
attributes under old used names (which are not really fitting naming 
convention of the resource attributes). So add a "aka" to the 
BaseComponent.
While we are here 
- repair some of the resource accesses from the unittests (i.e. direct 
conversion to dict), 
- update local dict storage, so that json.dumps works on update 
resource. Actually also fix the existing problem in the code with that.
- add _attributes_iterator to de-duplicate logic in Resource
- what a wonderful bug - passing sha256 as md5 (affects nodepool 
heavily). Add also tests for that.

Change-Id: I39edfc8fe4e4f6c216663c0e7602c9f1bec5cab4