Enable ansible module test for keypair to check return data
With a change in cloud layer to use proxy/resource layer for keypair
Ansible got borked and is not capable in resolving data we return it
back. Add test for it searching for the cause.
Investigation resulted in the fact, that when invoked by Ansible native
implementation of the Resoure.items was not suddifient under Py3, while
under Py2 it is fine. An attempt to integrate all invoked functions into SDK
from Ansible has not brought any success.

Change-Id: I60dec9ba26176efc5b8ad8378b0ef414754a857c