Fix discovery cache sharing
The intent is that connect_as shares the discovery cache so that
users don't pay the double cost. Unfortunately, that's not actually
what we were doing. There was also a mistaken thought that we should
be sharing a Session, but we can't do that, because the Session
has the Auth, and what we're doing is making a new Auth.

Attach the discovery_cache to the new Session directly and all
is good.

Change-Id: I07a362ea710b01d3588b06e9d3e0f1ce7eb101d4