Validate that connect_as connects as the project
We validate that a second keystone auth is made, but we don't
currently validate that the second auth is to the correct thing.

Add in a token payload validation, and have it set the project.

This will get subsummed in a bit with some work to make a proper
Fixture for the catalog/requests_mock stuff. But for now, this
should work.

Change-Id: I59c068c3a7e8e8d028b89da6a2f49e845d279473