Replace catalog-v3.json with keystoneauth fixture
As a first step to having a reusable fixture, replace our
use of catalog-v3.json with a fixture built on top of keystoneauth
token fixtures. Make a catalog containing all of the openstack
services.

There are a bunch of places where we just have some things
hardcoded that change with the auto-generation - like bare-metal
to baremetal. Those are whatever.

There are also places where the test code is requesting an admin
url but in the old catalog the admin and non-admin were
the same, so we were not catching that this was incorrect. Started
to fix this, but it got hairy and it's only a keystone v2 thing
which is on lifesupport. Make the admin and public urls the same
just for ease of landing this patch. We can go back in and make
them different and audit all the v2 keystone codepaths if anyone
decides to care about keystone v2.

There was a bunch of things hardcoded for block-storage which
break when we actually use proxies, which in turn breaks when
we try to properly use a better catalog fixture.

As a followup, we should stop forcing block_storage_api_version=2
because that's absurd, but this is a big enough patch as it is.

Also, volume quota calls were in the image file. Ooops.

Change-Id: I308cd159a5b71c94511f86c9d46bdbc589580c6d