Add header to auto-delete image upload objects
When using the tasks API, swift objects are used for creating the
final image. Although we manually clean these up when everything
goes right, it's possible when things go wrong that these objects
can be left lying around. Let's use the 'x-delete-after' header on
the objects to make sure they are cleaned up after a day.

Change-Id: Ieb2c11490a6fd195941ca87724eae2a6f67d43ad