Remove Accept header with empty value for HEAD and DELETE requests
According to RFC 2616, 7231 If server does not support requested content
type it should return 406 (not acceptable) RC. In case of HEAD or DELETE
we are not interested in the body of the response, but we should not
request missing content type. Either do not send this header at all, or
at least send "*/*" instead of empty value. Any project can start
following this spec at any time.

P.S. Due to some funny patch in my cloud Swift is rejecting all requests
(406) with "Accept: " header.

Change-Id: Id443bec2cbe80b3c3d1ad2a7208edac0dde4b21b