Properly convert baremetal fields to server-side values
Currently we pass the 'fields' argument as it is. If a caller provides
a client-side name instead of a server-side one (e.g. id instead of uuid),
it results in a failure. This patch fixes it.

Also stops requiring redundant 'name' for QueryParameters that are
specified as a dictionary (like fields).

Related-Bug: #1842989
Change-Id: Ie39067b6e9217eed63f5ae46331ab8a408f0ace6