Support uploading image from data and stdin
It might be useful to upload image from stdin. Disadvantage is, that it
is not possible to calculate checksums and this will prohibit threaded
image upload when using swift and tasks.
Additionally fix checksum validation during image creation - when
create_image is not called through cloud layer - the result might be
different, when checksums of existing image match (due to call and
return cloud.get_image). Fixing this requires also completing image v1
(proper find). Yeah, lots of tests are affected by that change.

Required-by: https://review.opendev.org/#/c/650374
Change-Id: I709d8b48cb7867fd806e2f19781bb84739363843