Bump min version of decorator
decorator.decorate decorator is only implemented in decorator 4.0.0.
The current lower bound of decorator is 3.4.0 but it does not work
with openstacksdk. It is found during a OSC change [1].
This commit bump the minimum version of decorator.
It looks like that decorator 4.0.0 works but we now test with
decorator 4.4.1 (the latest version),
so I chose the latest version as the minimum version.

[1] https://review.opendev.org/#/c/674325/

Change-Id: I7479a84ef19c7f1e881b9d56c12d2798340eb53b