Include "fields" to "SecurityGroup" query parameters
This new query parameter will allow to send a query sending the
"fields" parameter. This "fields" parameter contains the needed
OVO fields that need to be retrieved from the DB.

As commented in the related bug, the OS client "list" command only
prints five parameters, none of them the security group rules. In
systems with a reasonable amount of security groups, skipping the
unnecessary rule load can save a lot of time.

Change-Id: Ifa5af28e76bf6017fc6cd1dbd998efbe0e1200bf
Related-Bug: #1865223