Import generate_temp_url from swiftclient
The generate_temp_url function is used by other projects, for
example ironic, to get temporary urls that grant
unauthenticated access to the Swift object.
This patch simply takes the code from swiftclient, adapts
it for openstacksdk, and add it to the _proxy module in the
object_store section, to grant the use of the function directly
in openstacksdk.

Change-Id: I2de2047ef9025ac0bd191a973f2744b01b904ca7