Rationalize examples and functional extra config loading
This whole thing is a mess. examples is doing the right thing and
using get_extra_config, but its docs are wrong. functional is
reading config from a non-standard place but is the place that
the examples docs incorrect say it should be read from.

Update both places to read from the RIGHT location and update
the docs as well. Make examples support setting a key to use
and then set that key in tox so that we can just use the
functional settings for both.

Also, fix the clustering functional test so that it consumes
the IMAGE_NAME discovered in functional.base.

Change-Id: Ia9f150bbdd3825cb3e0d4282af3ca8f23dc1d888