Add bulk creation of rules for Security Group.
With new method _bulk_create in OpenStackSDK, now it is possible to use
it for API methods, which supports creating multiple instances in single
call. Neutron supports couple of resources to be created at once. In
this change we propose to add multiple rules creation for a single
security group.

Change-Id: I5855a6faa706adeaf738a797af0048985bffe65f