Revert "Switch to futurist for concurrency"
This partially reverts commit b6a22e3749555ecf9fea762dfc4935e312f6d8bb.

The addition of futurist as a hard depend broke python2, and at the
same time didn't mark the release as having done so. In looking at the
issue we realized that we don't actually need the hard depend.

Change-Id: I2d874f618f5b3f66d49cd2e964c6e05655f22c0f