Remove some unneeded things from test-requirements
We don't need to depend directly on python-subunit.
In python3 we don't need safe_hasattr.
We don't use testrepository.

In lower constraints, os-client-config and mox3 are
transitive depends that we do not care about so don't
need to track lower pins for them.

futures and ipaddress were removed from requirements
but not from lc.

Finally, remove future from lc and from __future__
since we're python3 only.

Change-Id: Ic22079f78f77221e83986e5a835457dbefc3ded7