Fix deleting stacks by id when waiting for result
When we want to wait for the result (wait=True), deleting a stack by
name works as expected. However, if deleting by id the stack is still
deleted but we get an error while polling events because no event is
ever found. This happens because the poller always compares the
resource_name attribute, which will never match the id.

This also required updating the tests as the fake events are always
created with the id as the resource_name and they assume that the
function is always called with the stack id and never the stack name.

Change-Id: Icd3a831faf14a45b1e492a0088160e519b5c6a02
Task: 30189
Story: 2005301