Add support for filters into the project cleanup
We might want do delete all resources i.e. older than X days. This is
really tricky, especially with regards to cleaning network service in
this case. To achieve this let's start looking to the port.device_id and
look, whether it will be present/deleted or not.
ATM only created_at and updated_at as filters are possible

Change-Id: I8b3f53f96f83dc0fbb09957097ed27755f255db1