Upload image via interop import if needed
We have support for interop import already, but it's not stitched
in to the image upload call. Add support for using it.

We're adding this as a fallback for the case that PUT is turned off
due to operator concerns about the need for scratch space. That
means we don't want existing users to all of a sudden start uploading
via import. However, if they do want to use import, we want that
to work, so there is now a flag.

Change-Id: I26e7ba5704d58a21f7ae2011e8c21e9b9310751a