placement: Add support for resource providers
We want some functionality in osc-placement. Add it here rather than
there.

Initially only the resource providers resource type is supported.
Additional resources will be added in future patches.

We're skipping the tests for the broken placement API versions, pending
discussions with keystone folks on how to resolve this long-term.

Change-Id: Ibf5f01b842e6fc79eb95c2c21bd69732de849597
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>