Switch nodepool test to containers
The nodepool-functional-container-openstack-siblings builds nodepool
container images (with openstacksdk and other "siblings" of nodepool
from source) and runs the build/boot tests with that.

The container builds of these tools are what OpenDev uses in
production, so this is more practical testing than installing by parts
on a bare Bionic system.  This allows us to remove the job duplication
in nodepool.

Change-Id: I91345ebf509238ca1f800cb595818c023cc56b0d