Add additional compute flavor operations
In order to proceed with switching flavor in OSC from novaclient/self
api onto SDK few new methods need to be supported.
Those include flavor access and set of extra_specs related operations.
While extra_specs in terms of SDK should be a separate resource, it is
not easy to implement it this way. On the other hand all those
operations are definive subset of flavor related operations.

In addition also merge FlavorDetail into Flavor class

Change-Id: Ia4f60acce5e0e5665e2ba704fe4cd0ec81437eb5