Include "fields" to "Port" query parameters
This new query parameter will allow to send a query with "fields"
parameter, that contains the needed OVO fields that need to be
retrieved from the BD.

As commented in the related bug, the OSC "list" command only prints
five parameters. Querying for only those ones will speed up the DB
operation and the dictionary creation.

Related-Bug: #1897100

Change-Id: I136a824896b4209c90ccc812a7e73064941e1b5a