Switch flavor ops in the cloud layer to proxy
Since very long time we want to switch cloud layer functions to use
underneath the proxy layer to reduce maintenance complexity. For flavors
we even did some tries in the past. This time it is meant even more
seriosly. Since flavors listing is a subject of caching, but
dogpile/pickle does not support caching of the complex objects returned
by funciton we convert and return flavors to Munch objects (we anyway
wanted to have it this way).

Change-Id: I0353bb8d1be69e18dd31f0abedf25818b42c14ce