Change nodepool job to build CentOS-8-stream (unblock gate)
Due to the recent renaming of the CentOS packages the job is currently
failing and requires fix in the diskimage-builder to be merged and
released. In the long run CentOS-8 as a standalone release is abandoned,
therefore verify building of the stream image instead.

Related-Bug: https://review.opendev.org/c/openstack/diskimage-builder/+/765963

Change-Id: I2aca89ce8bd27e920b0f80a66d6ed22d2277d05a