Allow unknown attributes to be included in request body
It is optional for resources to send in request's body attributes
which aren't included in mapping.
This is needed for Neutron to allow sending extra arguments in
POST/PUT requests to Neutron.

If that will be possible in SDK and OSC, Neutron team should be finally
able to deprecate (and drop) support for neutronclient CLI.

Change-Id: Ia3fd1052f438aedc6bc89bcca71834bcbafec9b9