Support SNAP_REAL_HOME when using openstacksdk inside a snap package
When the openstacksdk is used inside a snap package (eg. [1]), $HOME
is re-written by snapd so that each snap appears to have a dedicated
home directory that is a subdirectory of the real home directory [2].
But the clouds.yaml file is still in the real home directory.
To be able to use that file, let openstacksdk look for $SNAP_REAL_HOME
and if available, use that variable to construct the path for the
clouds.yaml file.

[1] https://opendev.org/x/snap-openstackclients
[2] https://snapcraft.io/docs/environment-variables

Change-Id: I570a98328188979f453bed98a2f8d6e35c569f19