Don't send empty remote_address_group_id for security groups
We are seeing an error raised by older Neutron's

 400: Client Error for url: <cloud>/v2.0/security-group-rules,
 Unrecognized attribute(s) 'remote_address_group_id'

This field was added unconditionally with
I50374c339ab7685a6e74f25f9521b8810c532e13 but, per above, appears to
cause problems for older Neutron instances.

To work around this, remove the argument from the body when blank by
overriding the _prepare_request function of SecurityGroupRule.

Two tests where this are used are updated; one checks the body is not
sent when None and the other is modified to send a
remote_address_group_id value to validate the other path.

Story: #2008577
Task: #41729
Change-Id: I25dabfde27b843df1c91c7fc37a1fe8d207b8010