Add min_count and max_count attributes to Server
The min_count and max_count kwargs are listed as valid kwargs on the
ComputeCloudMixin.create_server method but when they are specified,
they are not passed through to the eventual call to the nova REST API
because they are not defined as known/allowed attributes on the Server
class.

This adds min_count and max_count attributes to the Server class so
they will be honored when specified when calling create_server.

Story: 2008856
Task: 42377

Change-Id: Iaaedd292dc3dcf6486914d3b1f404ad95b12daeb