Drop cloud layer methods for Senlin
Rework of the cloud layer showed this was not working properly and tests
even verifying wrong things. Due to the big amount of effort required to
fix things and most likely nobody using it (this code is not really
usable now) drop it (comment-out). python-senlinclient is using direclty
proxy layer (correct), Ansible modules for senlin not existing - no
direct negative impact is visible.
If anybody is interested in getting it back - the code is there and can
be uncommented and fixed.

Change-Id: I198eed8454fd8ae2ce8fbd70e3562c9f15db7445