Switch Swift cloud layer to proxy
now it's turn of the object storage service to be relying on the proxy
functions in the cloud layer.
This includes not only completely relying on proxy implementation of the
functionality, but also extending proxy `get_object` method to be able
to save data to the file or to store it as `data` attribute of the
object instance.

Change-Id: Idc34738e42de7229a7c620b272c7385dde39eb1d