fix creation of protected image
Function _make_v2_image_params interpretes the value of the "protected"
key as a string when creating a new image, although it should
be a boolean. This causes a server-side schema validation
error (400 Bad Request) after sending the request body created
in this way.

This patch fixes the issue by adding "is_protected" to the
raw properties and also adds a unit test to validate correct behavior.

Story: 2009729
Task: 44141
Closes-Bug: https://bugs.launchpad.net/python-glanceclient/+bug/1927215
Signed-off-by: Jan Hartkopf <jhartkopf@inovex.de>
Change-Id: I3b318684ae8589acd688d0e724eca5703c25d85f