compute: Fix '*volume_attachment' proxy methods
The majority of these were totally broken and unusable because nova
doesn't allow you to retrieve a volume attachment by its ID: rather,
you need to use a combination of the server and volume IDs. As such,
each method should have been taking a combo of a Server object or ID and
a *Volume* object or ID, not a VolumeAttachment object or ID. These
fixes require rather extensive changes to the method signature, to the
point that we've actually added a release note to call this out. We've
included some hacky workarounds to avoid breaking existing users for
now, giving them time to migrate to the new, improved patterns.

In addition, the 'update_volume_attachment' proxy method never worked
since updates were forbidden on the resource (via 'allow_commit=False').
This is fixed by changing this value to True.

Change-Id: Iaab40ac4cc71f7626063700f476b53f7ffb5f39f
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>