image: Add "id" filter for images
The Glance v2 API supports a range of operators [1] for the list-images
API, allowing us to do things like:

  GET /v2/images?status=in:saving,queued

While there's no good reason to filter images by a single ID, it would
be helpful to be able to filter images by *multiple* IDs. Make this
possible by adding "id" to the list of possible filters.

[1] https://docs.openstack.org/api-ref/image/v2/index.html?expanded=list-images-detail#list-images

Change-Id: I735188e20d4e6c3b326e9080e878e717564b057d
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>