Improve StatsD metric precision
Make possible to see timings and counts for invoked APIs also based on
the returned status code. This helps to make deeper investigations of
the cloud behaviour (it might be very useful to see all variety of particular
status code occurence, especially error one). Since latency of 404 and
202 may vary heavily this also helps to reduce such interference.
In addition to that start emiting statsd metric for timeouts from API.

Change-Id: I8eb0174afc36f9ff10e2bd434f803f63736c160c