cloud: Update docstrings for network functions
We don't use munch objects here any more.

While we're here, we also fix a bug in the 'get_internal_networks' and
'get_external_networks' helpers: subclasses of 'Resource' including the
'Network' resource are not hashable so you can't put them in sets. We
must use lists and assume there are no duplicates.

Change-Id: I8cc7c5d83781698465c44d9a8137b76954de7f1d
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>