[stable-only] Pin openstacksdk in openstacksdk-functional-devstack
openstacksdk-functional-devstack jobs use master constraints and
also supposed to be run on stable branches. But sometimes master
constraints does not match with the versions present in old nodeset
in stable branches. For example, new constraints of 'coverage' 6.4
 I80d5c5bc2f379a9bac41e4bb40594a253853ee12
broke this job run on stable/ussuri. If we are using the master
constraints then it can break anytime when new constraints which
are not compatible with old branch supported distro.

Pinning this job to use stable/ussuri openstacksdk and constraints.

This squash the lower-constraints job removal fix
- https://review.opendev.org/c/openstack/openstacksdk/+/844096

NOTE(elod.illes): python2 binary is not needed in ussuri, besides
causing errors with requirements-check job. So binary is replaced
to python3 versions.

Story: 2010057
Tasks: 45501

Co-authored-by: Elod Illes <elod.illes@est.tech>

Change-Id: I11d60e42a83122b0e53b1d436ed4beca22bce5d4