Make nodepool jobs non voting
We ended up in dead loop with Zuul needed to pin older SDK version due
to some found issues, nodepool not able to cope with pin on its own and
sdk not able to merge any fixes for that. Make nodepool jobs non voting
for now unless we get out of problems.

Squash increasing of swap for devstack 
https://review.opendev.org/c/openstack/openstacksdk/+/844010 to get things
rolling

Change-Id: Ie8af2690f309df01022843a6fa934f0c548f32c3