resource: Fix pagination of nested Glance resources
This is a simple off-by-one error, but there are lots of things feeding
into this that I figured were worth explaining. Apologies in advance for
the length of this screed. Hopefully it's useful for someone.

There's a small bug in how we paginate when listing resources, or rather
when listing Glance objects. Glance doesn't return a 'Links' response
header or 'links' response body field like most other OpenStack
services. Instead, it returns a simple 'next' body field that is
relative. For example, consider the paginated response to the '/images'
API.

  {
    "images": [ ... ],
    "first":"/v2/images",
    "schema":"/v2/schemas/images",
    "next":"/v2/images?marker=823762fb-3128-43cf-8136-cee6b749f708"
  }

As a result, we're forced to treat Glance as yet another "special"
service when it comes to pagination. The first step of this is to look
for the 'next' field and grab the value. As you can see, this value
also includes the version prefix, which we don't want this since
keystoneauth1 already appends the version prefix when generating
absolute URLs from information in the service catalog. This means our
next step is to strip this version filter, which we have done since
change Iad8e69d20783cfe59c3e86bd6980da7dee802869. Unfortunately, the
original author of that change was a little too liberal in their
stripping (heh) and also removed the leading new slash. This causes
issues with the final part of our generation of the next link, namely
the deduplication of query parameters. Because most servers enforce an
upper limit on the size of URLs, it's important that we don't constantly
append our query parameters to the request URL for each page. Instead,
we parse the query parameters present in the 'next' URL and munge them
with our our parameters. The way we do this is using
'urllib.parse.urljoin' method with this pattern:

  urljoin(url, urlparse(url).path)

For example:

  >>> url = 'http://example.com/foo/bar?pet=dog&name=baxter'
  >>> urljoin(url, urlparse(url).path)
  'http://example.com/foo/bar'
  >>> url = '/foo/bar?pet=dog&name=baxter'
  >>> urljoin(url, urlparse(url).path)
  '/foo/bar'

Because we were stripping the leading '/', the first part of the path
was being identified as the server and was stripped.

  >>> url = 'foo/bar?pet=dog&name=baxter'
  >>> urljoin(url, urlparse(url).path)
  'foo/foo/bar'

After all that, the solution is simple. Stop stripping the leading
slash and generate a proper relative URL, as expected.

Change-Id: Ic9e1891ae7171a0cc45dd9faf1b2a6e37535b777
Co-authored-by: Stephen Finucane <stephenfin@redhat.com>
Story: 2010273
Task: 46200