Use /servers/detail endpoint in find_server proxy method
Currently, find_server will return the objects with different
fields filled in depending on whether the term provided is an id or a
name. This patch, in the same vein of [0], changes the find_server
method so it uses the /servers/detail endpoint to find servers in the
case of a name being provided. This ensures that the method fills in all
fields in every case.

[0] https://review.opendev.org/c/openstack/openstacksdk/+/854293

Change-Id: If8306f879361f872a1a87d589c756d9ae2435449