network: Remove duplicate module, update references
In change Id2513fd77fc303b42a52c436e3b0d46f93b7d376, we added support
for IPsec policies. However, in this change we appear to have
inadvertently merged two near-identical versions of the resource - one
in the 'vpn_ipsecpolicy' module and one in the 'vpn_ipsec_policy'
module. The latter is slightly more complete and the naming scheme
matches that of other modules so we keep that. We also opt to update the
proxy methods so that their naming scheme matches this pattern, along
with the docstrings for same.

Change-Id: I21799cf4ac6494f76b9c6b1963e214565a0e652d
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>