Initialize tests of real clouds
We want to start testing sdk and osc on real clouds. They are all
different, have different settings, policies, etc. In order to cope with
that it would be necessary to rework lot of tests and code to ensure all
of those "feautres" are properly detected by sdk and osc.

Block-storage test base class should not set operator cloud, only tests
that require it explicitly should use it

Change-Id: I09d118097bade8f0d8ff497957846a754ebc7c1e