Fix backup metadata management and update
max_microversion on the backup resource was missed causing no
microversion to be enabled at all for it. This leads to server refusing
managing metadata of the backup. Fix this and add corresponding test
with attempting to set metadata.
Since this change sets microversion we also need to ensure we add all
attributes of this version - encryption_key_id is the one added last.

Change-Id: I2c161ca30524bf230104958b8afdd8cf0b682a89