Introduce resource_registry in the proxies
Instead of hardcoding resource types on various places (what is
especially problematic once different api versions support different
resources) it would be possible to rely on a regisitry describing which
resources are supported by the certain proxy pointing to the exact
object classes. This is going to be used in a series of changes
implementing search_XXX method accepting string "service.resource" and
the method figuring out which exact class should be used as well as it
a first step into auto-generation of ansible modules.

Change-Id: If4abc89e88b0843877517cf9eaf5648778b325b0